Fixing a bug involving the returned changed region of the draw_line.
This fix along with another fix in the Framebuffer::refresh function.

Fixing a bug in the lua demo that was hidden by the two bug above.

Addressing these greatly introduced the refresh rate and dropped draw
latency. Now we are able to draw just like xochitl does, at least
as far as the eye can tell.