Update README to address compilation on recent systems
ARM toolchains on recent systems are often build with a newer Glibc than
the reMarkable has installed. This leads to runtime errors.

If the official toolchain from reMarkable is used, that's not an issue.

This commit updates the README with instructions on how to use this toolchain.
The link to the AUR package was removed as the most recent version also depends
on a too recent Glibc.

I've tested those steps on Debian testing.

Fixes #29.