Fix initial rendering
When an application that is using libremarkable is started right after
booting the device, it doesn't display anything.

In order to fix this problem, `pixclock` is set to 6250. This will result
in the expected flashing of the display as when starting other applications.
Though the rendering will be garbled/corrupted.

In order to fix that, the fix screen info needs to be read after the
var screen info was set. The fix screen info returns the number of lines
based on the rotation of the screen. Right after starting the reMarkable
the screen is set to landscape mode. libremarkable sets it to portrait mode
via the var screen info, which changes the line length of the fix screen info.

I also adjusted the `xres` and `yres` in the code to reflect the values
after the `ioctl` was successfully sent. Using the flipped values would
still work as they are automatically corrected, but using the right values
makes the code easier to understand.

I've tested it with the `demo`, both the traditional build as well as with
the musl build.

Fixes #32.