Point out that MXCFB_SEND_UPDATE is not ordinary
Is almost implemented a feature that would allow switching to
0x4048462e for updates.
When I read the mentioned documentation, I assumed that this
value would be an ordinal one since it seemed properly calculated
and therefore not "magic". Even though I trackd the iow! macro
down in ioctl-gen, I didn't quite understand it.
I compared the magic value with the "ordinary" just before committing
that new "feature".

Hopefully nobody will be dumb enough to re-add the same value as me,
but I added some documention just for documentations sake.