Remove gpio debouncing
The debouncing with a bitwise AND resulted in
Unpress events never being emitted.

Not sure but my guess would be that the demo.rs
never relied on Unpress events and it was therefore
overlooked.

I also don't see a reason for debouncing. The only
occurence might be for the first event if that's the
same as the intial gpio state. Every GPIO evdev event
sync immediately and there is not delay a debounce
applies to here (which is a typicial usecase of
debounce).