Client is not supposed to drop the ipc queue
Was an analog to the destroy() call on the c++ side.
Assumed this would be used by the client when closing,
but turns out, not the client neither the server ever
call this code. Seem to be an unused function that tripped
my when implementing it.

Dropping a framebuffer would cause the rm2fb stuff not to work
for any app until next reboot (effectively freezing the display).