Switch to using touch state instead of most recent event
Previously, the wacom code would emit WacomEvent::Hover iff the last
event seen was ToolPen. This has the wrong behaviour when the pen is
leaving the screen, since a Touch event was emitted as the pen lifts but
another ToolPen won't be emitted until the pen is no longer near.
Instead, we use the current touch state to decide if we're hovering or
not.

So why wasn't this always broken? In the ABS_DIST code farther down in
the file, it would update last_tool as part of some pressure / distance
corrections I don't fully understand. When we hit this codepath, the
resulting last-tool seems to be correct.  That code seems to reliably
fire on most tablets, but it seems that it's not guaranteed to happen.

Anyways this is all a bit sketchy, and we should test on a couple
devices before calling it good.