Fix submodule URL for Freetype
Due to recent [changes in GitHub’s supported protocols](https://github.blog/2021-09-01-improving-git-protocol-security-github/), the fetching of submodules of this repository is failing. This also causes [folly](https://github.com/bkirwi/folly)’s build to fail, since for some reason the submodules are fetched as part of the build (@bkirwi).

This patch simply updates the submodule’s URL to use https:// instead of git://.